VEGAS+: Avoid calculations with nan values
Numpy shows warnings even if the nan values are removed after their calculation. If there are no intentional calculations with nan values, these warnings can be helpful to detect bugs.
The acc variable in _warmup_grid and _run_iteration is only used to inform the user in INFO and DEBUG logs; now it no longer shows nan for negative or zero-valued integrands.