Change the VEGASMap update behaviour when the integrand wasn't evaluated in all intervals
If by chance no sample fell into some interval, the weight in this interval is now set to the weight of the neighbouring interval instead of zero.
Intervals where no samples fall into appear especially in VEGAS._warmup_grid, which uses a small number of points per iteration and no stratification. Increasing N does not help because the number of intervals per dimension depends on N.
The VEGAS quadrature should be a bit more accurate now, especially with constant integrands.

Negative values in sig2 very close to 0 appear now in _run_iteration, too, and not only in _warmup_grid when running the tests.